+++ xxcmd/dbitem.py
@@ -21,13 +21,3 @@ class DBItem():
     # Return a string suitable for substring searching
     def search_key(self):
         return "{0} {1}".format(self.label, self.cmd).lower()
-
-    # Return a pretty string representation
-    def pretty(self, indent=0, show_label=True):
-
-        if self.label and show_label:
-            label = self.label.ljust(indent) + " "
-        else:
-            label = ""
-
-        return "{0}{1}".format(label, self.cmd)